重构结构 重构文件存储服务，规范一部分命名
