+++ WordCountMain/src/main/java/priv/dawn/wordcount/WordCountMainApplication.java
@@ -1,0 +1,15 @@
+package priv.dawn.wordcount;
+
+import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@EnableDubbo
+@SpringBootApplication
+public class WordCountMainApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(WordCountMainApplication.class, args);
+    }
+
+}
