+++ WordCountMain/src/main/java/priv/dawn/wordcount/config/WebSocketConfiguration.java
@@ -1,0 +1,15 @@
+package priv.dawn.wordcount.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.socket.server.standard.ServerEndpointExporter;
+
+@Configuration
+public class WebSocketConfiguration {
+
+    @Bean
+    public ServerEndpointExporter serverEndpointExporter() {
+        return new ServerEndpointExporter();
+    }
+
+}
