+++ WordCountMain/src/main/java/priv/dawn/wordcount/pojo/dto/FileChunkDTO.java
@@ -1,0 +1,11 @@
+package priv.dawn.wordcount.pojo.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class FileChunkDTO {
+    private int chunkId;
+    private String context;
+}
