+++ WordCountMain/src/main/java/priv/dawn/wordcount/pojo/vo/TextFileVo.java
@@ -1,0 +1,11 @@
+package priv.dawn.wordcount.pojo.vo;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class TextFileVo {
+    private String filename;
+    private String context;
+}
