+++ WordCountMain/src/main/java/priv/dawn/wordcount/service/FileStorageService.java
@@ -1,0 +1,8 @@
+package priv.dawn.wordcount.service;
+
+import priv.dawn.wordcount.pojo.vo.TextFileVo;
+
+public interface FileStorageService {
+    int saveFile(TextFileVo fileText);
+    TextFileVo getFile(int fileUID);
+}
