+++ WordCountMain/src/main/java/priv/dawn/wordcount/service/WordCountService.java
@@ -1,0 +1,9 @@
+package priv.dawn.wordcount.service;
+
+import priv.dawn.wordcount.domain.WordCountStateEnum;
+import priv.dawn.wordcount.pojo.vo.WordCountListVO;
+
+public interface WordCountService {
+    WordCountStateEnum startCountWord(int fileUID);
+    WordCountListVO getWordCounts(int fileUID);
+}
