--- WordCountMain/src/main/java/priv/dawn/wordcountmain/WordCountMainApplication.java
@@ -1,15 +1,0 @@
-package priv.dawn.wordcountmain;
-
-import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@EnableDubbo
-@SpringBootApplication
-public class WordCountMainApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(WordCountMainApplication.class, args);
-    }
-
-}
