--- WordCountMain/src/main/java/priv/dawn/wordcountmain/config/WebSocketConfiguration.java
@@ -1,15 +1,0 @@
-package priv.dawn.wordcountmain.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.socket.server.standard.ServerEndpointExporter;
-
-@Configuration
-public class WebSocketConfiguration {
-
-    @Bean
-    public ServerEndpointExporter serverEndpointExporter() {
-        return new ServerEndpointExporter();
-    }
-
-}
