--- WordCountMain/src/main/java/priv/dawn/wordcountmain/pojo/dto/FileChunkDTO.java
@@ -1,11 +1,0 @@
-package priv.dawn.wordcountmain.pojo.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class FileChunkDTO {
-    private int chunkId;
-    private String context;
-}
