--- WordCountMain/src/main/java/priv/dawn/wordcountmain/service/FileStorageService.java
@@ -1,8 +1,0 @@
-package priv.dawn.wordcountmain.service;
-
-import priv.dawn.wordcountmain.pojo.vo.FileTextVO;
-
-public interface FileStorageService {
-    int saveFile(FileTextVO fileText);
-    String getContext(int fileUID);
-}
