--- WordCountMain/src/main/java/priv/dawn/wordcountmain/service/WordCountService.java
@@ -1,9 +1,0 @@
-package priv.dawn.wordcountmain.service;
-
-import priv.dawn.wordcountmain.domain.WordCountStateEnum;
-import priv.dawn.wordcountmain.pojo.vo.WordCountListVO;
-
-public interface WordCountService {
-    WordCountStateEnum startCountWord(int fileUID);
-    WordCountListVO getWordCounts(int fileUID);
-}
