+++ WordCountMain/src/test/java/priv/dawn/wordcount/WordCountMainApplicationTests.java
@@ -1,0 +1,16 @@
+package priv.dawn.wordcount;
+
+import org.junit.jupiter.api.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+class WordCountMainApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
