--- WordCountMain/src/test/java/priv/dawn/wordcountmain/WordCountMainApplicationTests.java
@@ -1,16 +1,0 @@
-package priv.dawn.wordcountmain;
-
-import org.junit.jupiter.api.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-class WordCountMainApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
