+++ pom.xml
@@ -27,7 +27,6 @@
         <maven.compiler.target>8</maven.compiler.target>
         <!--版本控制以后不写在这-->
         <spring-boot.version>2.6.13</spring-boot.version>
-        <mybatis-generator.version>1.3.2</mybatis-generator.version>
         <dubbo.version>3.0.9</dubbo.version>
         <nacos.version>2.3.2</nacos.version>
     </properties>
@@ -133,33 +132,6 @@
                             </goals>
                         </execution>
                     </executions>
-                </plugin>
-                <plugin>
-                    <groupId>org.mybatis.generator</groupId>
-                    <artifactId>mybatis-generator-maven-plugin</artifactId>
-                    <version>${mybatis-generator.version}</version>
-<!--                    <configuration>-->
-<!--                        <configurationFile>src/main/resources/mybatis-generator/generatorConfig.xml-->
-<!--                            &lt;!&ndash;                            TODO 这里需要改一下每个项目自己的配置-->
-<!--                        </configurationFile>-->
-<!--                        <verbose>true</verbose>-->
-<!--                        <overwrite>true</overwrite>-->
-<!--                    </configuration>-->
-                    <executions>
-                        <execution>
-                            <id>Generate MyBatis Artifacts</id>
-                            <goals>
-                                <goal>generate</goal>
-                            </goals>
-                        </execution>
-                    </executions>
-<!--                    <dependencies>-->
-<!--                        <dependency>-->
-<!--                            <groupId>org.mybatis.generator</groupId>-->
-<!--                            <artifactId>mybatis-generator-core</artifactId>-->
-<!--                            <version>${mybatis-generator.version}</version>-->
-<!--                        </dependency>-->
-<!--                    </dependencies>-->
                 </plugin>
             </plugins>
         </pluginManagement>
