--- word-count-api/src/main/java/priv/dawn/wordcount/api/Worker.java
@@ -1,14 +1,0 @@
-package priv.dawn.wordcount.api;
-
-/**
- * Created with IntelliJ IDEA.
- *
- * @Description:
- * @Auther: Dawn Yang
- * @Since: 2024/10/09/11:05
- */
-public interface Worker {
-
-
-
-}
