+++ word-count-api/src/main/java/priv/dawn/wordcount/api/WorkerService.java
@@ -1,0 +1,12 @@
+package priv.dawn.wordcount.api;
+
+/**
+ * Created with IntelliJ IDEA.
+ *
+ * @Description:
+ * @Auther: Dawn Yang
+ * @Since: 2024/10/09/11:05
+ */
+
+public interface WorkerService {
+}
