+++ word-count-api/src/main/java/priv/dawn/wordcount/domain/ChunkVo.java
@@ -1,0 +1,19 @@
+package priv.dawn.wordcount.domain;
+
+import lombok.Data;
+
+/**
+ * Created with IntelliJ IDEA.
+ *
+ * @Description: 文件分区 VO
+ * @Auther: Dawn Yang
+ * @Since: 2024/10/12/21:28
+ */
+
+@Data
+public class ChunkVo {
+
+    private Integer chunkId;
+    private String context;
+
+}
